SDIT-743 Refactor to individual services
